Separate the test cases.
